+++ libraries/libldap/schema.c
@@ -2341,7 +2341,7 @@ static char *const err2text[] = {
 char *
 ldap_scherr2str(int code)
 {
-	if ( code < 0 || code >= (sizeof(err2text)/sizeof(char *)) ) {
+	if ( code < 0 || code >= (int)(sizeof(err2text)/sizeof(char *)) ) {
 		return "Unknown error";
 	} else {
 		return err2text[code];
